SG-7625: Fixes missing slashes when a windows folder starts with an underscore (#2)
Note: This is an issue with the report sent back to the browser for the user to see when they've run folder creation from browser integration, or when looking at logs. The app functioned properly in this situation, otherwise.

The backslash escapes the underscore so paths such as `c:\asset\_psd\` renders as `c:\asset_psd\` which can be confusing when setting up templates.